Single large Word database, theme filtering
In this approach, all words are imported into a single 'Word' database,
and a smaller list of words to randomly sample from is generated by
simultaneously filtering for first letter, part of speech, and theme.
This is less efficient than having separate pre-organized tables,
however is much easier to manage (for the time being).