Handle duplicate words in acrostic
Modified generate.py to filter the list of available words based on
duplicate entries - in the updated behavior, if a word is already in
use in the acrostic, it is not selected again.  However, if all
possible post-filtered words are already in use, then among this list
one is chosen at random to duplicate.