stub: pass current word list to constructions
Created stub methods demonstrating how to pass the current word list to
the build_constructions module.  Instead of building up all of the
filters in one step and sending them back, this module now builds a set
of filters for each character in the word, based on both the position
and the current list of words.  This allows for more flexible
constructions (for example, create an adverb at position 3 if position
2 is a noun, etc etc)