Changed construction procedure
Instead of deleting an entire category, the likelihood for that
category is reduced and put back in the pool for random sampling.  This
matters now that we use a more flexible verb-generating protocol, where
we are more likely to hit verb-subject acrostics than before.  Also
slightly modified proportions of elements.